requirements.txt
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.26
django-filter==1.1.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyMySQL==0.9.3
pytz==2019.3
requests==2.22.0
six==1.13.0
uritemplate==3.0.0
urllib3==1.25.7
